Move around some vehicle & prefab code
